move setup of .corgidrp dirs out of setup
