+++ .github/workflows/python-app.yml
@@ -28,7 +28,6 @@ jobs:
         python -m pip install --upgrade pip
         pip install flake8 pytest flake8-docstrings-complete
         if [ -f requirements.txt ]; then pip install -r requirements.txt -e .; fi
-        cat ~/.corgidrp/corgidrp.cfg
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
