+++ corgidrp/caldb.py
@@ -285,3 +285,13 @@ class CalDB:
         # load all these files into the caldb
         for calib_frame in calib_frames:
             self.create_entry(calib_frame, to_disk=to_disk)
+
+### Create set of default calibrations
+# Add default detector_params calibration file if it doesn't exist
+if not os.path.exists(os.path.join(corgidrp.default_cal_dir, "DetectorParams_2023-11-01T00:00:00.000.fits")):
+    default_detparams = data.DetectorParams({}, date_valid=time.Time("2023-11-01 00:00:00", scale='utc'))
+    default_detparams.save(filedir=corgidrp.default_cal_dir)
+
+    # add default caldb entries
+    default_caldb = CalDB()
+    default_caldb.scan_dir_for_new_entries(corgidrp.default_cal_dir)
